Support .ccls files in subdirectories even without one at the project root
Current setup of ccls - A .ccls file needs to be at the project root and then
.ccls files in sub directories can be used to add compiler flags for specific folders.

With this change, when we open a file, we recursively look for a .ccls file starting from
the current directory up the directory tree for a .ccls file. We do not assume that a .ccls
file at the parent is going to provide any inputs to the compiler flags for a file.